[SPARK-39808][SQL] Support aggregate function MODE
