+++ sql/core/src/test/resources/sql-tests/results/group-by.sql.out
@@ -845,3 +845,22 @@ struct<a:int,collect_list(b):array<int>,collect_list(b):array<int>>
 -- !query output
 1	[4,4]	[4,4]
 2	[3,4]	[3,4]
+
+
+-- !query
+SELECT mode(a), mode(b) FROM testData
+-- !query schema
+struct<mode(a):int,mode(b):int>
+-- !query output
+3	1
+
+
+-- !query
+SELECT a, mode(b) FROM testData GROUP BY a ORDER BY a
+-- !query schema
+struct<a:int,mode(b):int>
+-- !query output
+NULL	1
+1	1
+2	1
+3	1
